sprint_1034_46

+++ system/module_core/backend/patterns/Module.php
@@ -19,18 +19,21 @@ namespace effcore {
 
   function uninstall() {
   # reverse the deployment process: delete files
-    foreach (storage::get('files')->select('deployment/'.$this->id.'/copy') ?? [] as $c_info) {
-      $c_file = new file($c_info->to);
-      if (@unlink($c_file->path_get()))
-           message::insert(new text('File "%%_path" was deleted.',     ['path' => $c_file->path_get_relative()]));
-      else message::insert(new text('File "%%_path" was not deleted!', ['path' => $c_file->path_get_relative()]), 'error');
+    $deployment = storage::get('files')->select('deployment');
+    if ( isset($deployment[$this->id]['copy']) ) {
+      foreach ($deployment[$this->id]['copy'] as $c_info) {
+        $c_file = new file($c_info->to);
+        if (@unlink($c_file->path_get()))
+             message::insert(new text('File "%%_path" was deleted.',     ['path' => $c_file->path_get_relative()]));
+        else message::insert(new text('File "%%_path" was not deleted!', ['path' => $c_file->path_get_relative()]), 'error');
+      }
     }
-  # delete instances
+  # reverse the deployment process: delete instances
     foreach (instance::get_all_by_module($this->id) as $c_instance) {
       if ($c_instance->select())
           $c_instance->delete();
     }
-  # delete entities
+  # reverse the deployment process: delete entities
     foreach (entity::get_all_by_module($this->id) as $c_entity) {
       if ($c_entity->uninstall())
            message::insert(new text('Entity "%%_entity" was uninstalled.',     ['entity' => $c_entity->name])         );
